% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proharmed.R
\name{count_intersections}
\alias{count_intersections}
\title{Title Count Intersections}
\usage{
count_intersections(data, threshold = 1)
}
\arguments{
\item{data}{}

\item{threshold}{}
}
\description{
Title Count Intersections
}
