% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{create_filter_detailed_plot}
\alias{create_filter_detailed_plot}
\title{Detailed plot for logging data of mqhandler's filter_protein_ids method}
\usage{
create_filter_detailed_plot(
  logging,
  organism,
  reviewed,
  decoy,
  out_dir = NULL,
  file_type = "png"
)
}
\arguments{
\item{logging}{detailed logging Dataframe that has been returned by the filtered_protein_ids method of mqhandler}

\item{organism}{organism that has been specified in the filter_protein_ids method of mqhandler}

\item{reviewed}{specified reviewed parameter of the filter_protein_ids method of mqhandler}

\item{decoy}{specified rev_con parameter of the filter_protein_ids method of mqhandler}

\item{out_dir}{(optional) output directory to save the plot}

\item{file_type}{(optional) file type for the plot}
}
\value{
ggplot object
}
\description{
Detailed plot for logging data of mqhandler's filter_protein_ids method
}
