% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{create_ortholog_detailed_plot}
\alias{create_ortholog_detailed_plot}
\title{Detailed plot for logging data of mqhandler's map_orthologs method}
\usage{
create_ortholog_detailed_plot(
  logging,
  organism,
  out_dir = NULL,
  file_type = "png"
)
}
\arguments{
\item{logging}{detailed logging Dataframe that has been returned by the map_orthologs method of mqhandler}

\item{organism}{organism that has been specified in the map_orthologs method of mqhandler}

\item{out_dir}{(optional) output directory to save the plot}

\item{file_type}{(optional) file type for the plot}
}
\description{
Detailed plot for logging data of mqhandler's map_orthologs method
}
