% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{create_overview_plot}
\alias{create_overview_plot}
\title{Overview plot for logging data of mqhandler's functions}
\usage{
create_overview_plot(logging, out_dir = NULL, file_type = "png")
}
\arguments{
\item{logging}{overview logging Dataframe that has been returned by one of the four methods of mqhandler}

\item{out_dir}{(optional) output directory to save the plots}

\item{file_type}{(optional) file type for the plots (png, pdf, jpg, ...)}
}
\value{
list of two ggplot objects
}
\description{
Overview plot for logging data of mqhandler's functions
}
