% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{create_reduced_detailed_plot}
\alias{create_reduced_detailed_plot}
\title{Detailed plot for logging data of mqhandler's reduce_genenames method}
\usage{
create_reduced_detailed_plot(logging, out_dir = NULL, file_type = "png")
}
\arguments{
\item{logging}{detailed logging Dataframe that has been returned by the reduce_genenames method of mqhandler}

\item{out_dir}{(optional) output directory to save the plot}

\item{file_type}{(optional) file type for the plot}
}
\value{
ggplot object
}
\description{
Detailed plot for logging data of mqhandler's reduce_genenames method
}
