% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mqhandler.R
\name{filter_protein_ids}
\alias{filter_protein_ids}
\title{Filter Protein IDs

Remove decoy IDs, contaminated IDs and / or filter IDs by organism.}
\usage{
filter_protein_ids(
  data,
  protein_column,
  organism = NULL,
  decoy = FALSE,
  action = "delete",
  gene_column = NULL,
  reviewed = TRUE
)
}
\arguments{
\item{data}{a Dataframe with a column containing the protein IDs}

\item{protein_column}{name of column with protein IDs}

\item{organism}{(optional) specify organism the ids should match to}

\item{decoy}{(optional) bool to indicate if protein ids from decoy fasta (REV__, CON__) should be kept}

\item{action}{(optional) what to do, if IDs cell is empty after filtering. Keep empty cell, delete it or fill it based on gene name.}

\item{gene_column}{(optional) name of column with gene names}

\item{reviewed}{(optional) bool to indicate if newly retrieved protein IDs should be reduced to reviewed ones}
}
\value{
filtered Dataframe
}
\description{
Filter Protein IDs

Remove decoy IDs, contaminated IDs and / or filter IDs by organism.
}
