% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mqhandler.R
\name{get_orthologs}
\alias{get_orthologs}
\title{Get Orthologs

Get ortholog gene names from origin organism to target organism.}
\usage{
get_orthologs(data, gene_column, organism, tar_organism)
}
\arguments{
\item{data}{a Dataframe with a column containing the gene names}

\item{gene_column}{name of column with gene names}

\item{organism}{specify organism the ids currently match to}

\item{tar_organism}{specify organism the ids should match to}
}
\value{
dataframe with orthologs
}
\description{
Get Orthologs

Get ortholog gene names from origin organism to target organism.
}
