% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mqhandler.R
\name{reduce_genenames}
\alias{reduce_genenames}
\title{Reduce Gene Names

Use existing Gene Names to reduce them and replace them by a single Name 
to redundant Names. Following modes are possible.

Modes:
ensembl - Use gProfiler to reduce gene names to those having a Ensembl ID
HGNC - Use HGNC database to reduce gene names to those having an entry in HGNC (only for human)
mygeneinfo - Use mygeneinfo database to reduce gene names to those having an entry in mygeneinfo
enrichment - Use gProfiler to reduce gene names to those having a functional annotation}
\usage{
reduce_genenames(
  data,
  mode,
  gene_column,
  organism,
  res_column = NULL,
  keep_empty = FALSE,
  HGNC_mode = "mostfrequent"
)
}
\arguments{
\item{data}{a Dataframe with a column containing the protein IDs}

\item{mode}{mode of refilling (See in description for more infos)}

\item{gene_column}{name of column with gene names}

\item{organism}{(optional) specify organism the ids should match to}

\item{res_column}{name of column of reduced gene names results. If NULL, the gene_column will be overridden}

\item{keep_empty}{bool to indicate if empty reduced gene names cells should be kept or deleted}

\item{HGNC_mode}{setting on how to selected the gene names in HGNC (mostfrequent, all)}
}
\value{
reduced Dataframe
}
\description{
Reduce Gene Names

Use existing Gene Names to reduce them and replace them by a single Name 
to redundant Names. Following modes are possible.

Modes:
ensembl - Use gProfiler to reduce gene names to those having a Ensembl ID
HGNC - Use HGNC database to reduce gene names to those having an entry in HGNC (only for human)
mygeneinfo - Use mygeneinfo database to reduce gene names to those having an entry in mygeneinfo
enrichment - Use gProfiler to reduce gene names to those having a functional annotation
}
