% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mqhandler.R
\name{filter_protein_ids}
\alias{filter_protein_ids}
\title{Filter Protein IDs

Remove decoy IDs, contaminated IDs and / or filter IDs by organism.}
\usage{
filter_protein_ids(
  data,
  protein_column,
  organism = NULL,
  decoy = FALSE,
  keep_empty = FALSE,
  res_column = NULL,
  reviewed = TRUE
)
}
\arguments{
\item{data}{a Dataframe with a column containing the protein IDs}

\item{protein_column}{name of column with protein IDs}

\item{organism}{(optional) specify organism the ids should match to}

\item{decoy}{(optional) bool to indicate if protein ids from decoy fasta (REV__, CON__) should be kept}

\item{keep_empty}{bool to indicate if empty reduced gene names cells should be kept or deleted}

\item{res_column}{name of column of filtered protein IDs. If NULL, the protein_column will be overridden}

\item{reviewed}{(optional) bool to indicate if newly retrieved protein IDs should be reduced to reviewed ones}
}
\value{
filtered Dataframe
}
\description{
Filter Protein IDs

Remove decoy IDs, contaminated IDs and / or filter IDs by organism.
}
