% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mqhandler.R
\name{filter_protein_ids}
\alias{filter_protein_ids}
\title{Title}
\usage{
filter_protein_ids(
  data,
  protein_column,
  organism = NULL,
  decoy = FALSE,
  action = "delete",
  gene_column = NULL,
  reviewed = TRUE
)
}
\arguments{
\item{data}{}

\item{protein_column}{name of column with protein IDs
Optional:}

\item{organism}{specify organism the ids should match to}

\item{decoy}{}

\item{action}{what to do, if IDs cell is empty after filtering. Keep empty cell, delete it or fill it based on gene name.}

\item{gene_column}{name of column with gene names}

\item{reviewed}{}
}
\value{

}
\description{
Title
}
