% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proharmed.R
\name{remap_genenames}
\alias{remap_genenames}
\title{Remap Gene Names}
\usage{
remap_genenames(
  data,
  mode,
  protein_column,
  gene_column = "Gene Names",
  res_column = NULL,
  skip_filled = FALSE,
  organism = NULL,
  fasta = NULL,
  keep_empty = FALSE
)
}
\arguments{
\item{data}{a Dataframe with a column containing the protein IDs}

\item{mode}{mode of refilling (See in description for more infos)}

\item{protein_column}{name of column with protein IDs}

\item{gene_column}{(optional) name of column with gene names}

\item{res_column}{name of column of reduced gene names results. If NULL, the gene_column will be overridden}

\item{organism}{(optional) specify organism the ids should match to}

\item{fasta}{(optional) fasta file when mode all or fasta}

\item{keep_empty}{bool to indicate if empty reduced gene names cells should be kept or deleted}
}
\value{
remapped Dataframe
}
\description{
Use protein IDs to get associated Gene Names and fill exmpty entries 
and optionally replace already existing Names. Following modes are possible.

Modes:
all - Use primarly fasta infos and additionally uniprot infos.
fasta - Use information extracted from fasta headers.
uniprot - Use mapping information from uniprot and use all gene names.
uniprot_primary - Use mapping information from uniprot and only all primary gene names.
uniprot_one - Use mapping information from uniprot and only use most frequent single gene name.
}
